wswan: Fix AAM/AAD opcode implementation.
Based on the available V30MZ documentation and hardware verification, the V30MZ
(like the 80186, unlike the V20/V30) respects the immediate argument provided
to AAM and AAD opcodes. Most WonderSwan emulators get this wrong, but emulating
this correctly is required to enable certain modern C compiler optimizations
for homebrew.